// Employee Management System/src/app/models/employee.model.ts
export class Employee {
  id: number = 0;
  FirstName: string = '';
  LastName: string = '';
  Email: string = '';
  PhoneNumber: string = '';
  Department: string = '';
  JobTitle: string = '';
  HireDate: string = '';
  TerminationDate: string | null = '';
  IsActive: boolean = true;
  CreatedAt: string = '';
  UpdatedAt: string = '';
  photo: string = '';
  appraisal: string = '';
}